refactoring test cases for events
