+++ crates/client/rpc/src/events/mod.rs
@@ -1,6 +1,10 @@
+#[cfg(test)]
+mod tests;
+
 use std::iter::Skip;
 use std::vec::IntoIter;
 
+use jsonrpsee::core::RpcResult;
 use log::error;
 use mc_rpc_core::utils::get_block_by_block_hash;
 use mp_starknet::block::Block;
@@ -13,11 +17,12 @@ use sc_client_api::backend::{Backend, StorageProvider};
 use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 use sp_runtime::traits::Block as BlockT;
-use starknet_core::types::BlockId;
+use starknet_core::types::{BlockId, EventsPage};
 use starknet_ff::FieldElement;
 
 use crate::errors::StarknetRpcApiError;
-use crate::Starknet;
+use crate::types::RpcEventFilter;
+use crate::{EmittedEvent, Starknet};
 
 impl<B, BE, C, P, H> Starknet<B, BE, C, P, H>
 where
@@ -58,6 +63,82 @@ where
                 Vec::new()
             });
         Ok((block_events, block))
+    }
+
+    /// Helper function to filter Starknet events provided a RPC event filter
+    ///
+    /// # Arguments
+    ///
+    /// * `filter` - The RPC event filter
+    ///
+    /// # Returns
+    ///
+    /// * `EventsPage` - The filtered events with continuation token
+    pub fn filter_events(&self, filter: RpcEventFilter) -> RpcResult<EventsPage> {
+        let mut filtered_events = vec![];
+        let mut index = 0;
+
+        // get filter values
+        let mut current_block = filter.from_block;
+        let to_block = filter.to_block;
+        let from_address = filter.from_address;
+        let keys = filter.keys;
+        let mut continuation_token = filter.continuation_token;
+        let chunk_size = filter.chunk_size;
+
+        // Iterate on block range
+        while current_block <= to_block {
+            let (block_events, block) = self.get_block_events(current_block)?;
+            let block_events_len = block_events.len();
+            // if block_events length < continuation_token, keep going and reduce the pagination
+            if block_events_len < continuation_token {
+                continuation_token -= block_events_len;
+                index += block_events_len;
+                current_block += 1;
+                continue;
+            }
+
+            let block_events = block_events.into_iter().skip(continuation_token);
+            // Kept in order to calculate continuation token.
+            let block_events_len = block_events_len - continuation_token;
+            let index_before_loop = index;
+
+            let block_hash = block.header().hash(*self.hasher).into();
+            let block_number = block.header().block_number.try_into().map_err(|e| {
+                error!("Failed to convert block number to u64: {e}");
+                StarknetRpcApiError::BlockNotFound
+            })?;
+
+            let (new_filtered_events, continuation_index) = filter_events_by_params(
+                block_events,
+                from_address,
+                keys.clone(),
+                Some((chunk_size as usize) - filtered_events.len()),
+            );
+            index = continuation_index;
+
+            filtered_events.extend(
+                new_filtered_events
+                    .iter()
+                    .map(|event| EmittedEvent {
+                        from_address: event.from_address.into(),
+                        keys: event.keys.clone().into_iter().map(|key| key.into()).collect(),
+                        data: event.data.clone().into_iter().map(|data| data.into()).collect(),
+                        block_hash,
+                        block_number,
+                        transaction_hash: event.transaction_hash.into(),
+                    })
+                    .collect::<Vec<EmittedEvent>>(),
+            );
+
+            if filtered_events.len() >= chunk_size as usize {
+                let token = if index - index_before_loop < block_events_len { Some((index).to_string()) } else { None };
+                return Ok(EventsPage { events: filtered_events, continuation_token: token });
+            }
+            current_block += 1;
+            continuation_token = 0;
+        }
+        Ok(EventsPage { events: filtered_events, continuation_token: None })
     }
 }
