+++ crates/client/rpc/src/events/tests.rs
